Update Downloader.cs
Expand Message-ID Regex to include pipes and double quotes. This was to address an bug where message IDs had a pipe char in, Downloader then failed when trying to create a file with a pipe in the name. I think the ideal approach would be similar to the "illegalChars" one or one based around encoding URL-style, but I couldn't get that to work quickly and this solved the immediate problem. Not sure if the double quote is of practical value but included for completeness.